tests/test_performance_optimizer.py
import pytest
import numpy as np
from unittest.mock import MagicMock
from src.opentts_tensorflow_integration.core import PerformanceOptimizer, BaseTTS

class MockTTS(BaseTTS):
    def __init__(self):
        super().__init__()
        self.current_voice = "default"

    def synthesize(self, text):
        return np.array([0.1, 0.2, 0.3])

    def load_model(self, model_path):
        pass

    @property
    def available_voices(self):
        return ["default"]

    def set_voice(self, voice_name):
        self.current_voice = voice_name

@pytest.fixture
def performance_optimizer():
    return PerformanceOptimizer(cache_size=2)

def test_cache_audio(performance_optimizer):
    audio = np.array([0.1, 0.2, 0.3])
    performance_optimizer.cache_audio("Hello", "MockTTS", "default", audio)
    cached_audio = performance_optimizer.get_cached_audio("Hello", "MockTTS", "default")
    np.testing.assert_array_equal(audio, cached_audio)

def test_cache_size_limit(performance_optimizer):
    audio1 = np.array([0.1, 0.2, 0.3])
    audio2 = np.array([0.4, 0.5, 0.6])
    audio3 = np.array([0.7, 0.8, 0.9])

    performance_optimizer.cache_audio("Hello1", "MockTTS", "default", audio1)
    performance_optimizer.cache_audio("Hello2", "MockTTS", "default", audio2)
    performance_optimizer.cache_audio("Hello3", "MockTTS", "default", audio3)

    assert performance_optimizer.get_cached_audio("Hello1", "MockTTS", "default") is None
    np.testing.assert_array_equal(audio2, performance_optimizer.get_cached_audio("Hello2", "MockTTS", "default"))
    np.testing.assert_array_equal(audio3, performance_optimizer.get_cached_audio("Hello3", "MockTTS", "default"))

def test_batch_process(performance_optimizer):
    tts_engine = MockTTS()
    texts = ["Hello", "World", "Test"]
    results = performance_optimizer.batch_process(tts_engine, texts)

    assert len(results) == 3
    for result in results:
        np.testing.assert_array_equal(result, np.array([0.1, 0.2, 0.3]))

def test_batch_process_with_cache(performance_optimizer):
    tts_engine = MockTTS()
    cached_audio = np.array([0.4, 0.5, 0.6])
    performance_optimizer.cache_audio("Cached", "MockTTS", "default", cached_audio)

    texts = ["Cached", "New"]
    results = performance_optimizer.batch_process(tts_engine, texts)

    assert len(results) == 2
    np.testing.assert_array_equal(results[0], cached_audio)
    np.testing.assert_array_equal(results[1], np.array([0.1, 0.2, 0.3]))

def test_clear_cache(performance_optimizer):
    audio = np.array([0.1, 0.2, 0.3])
    performance_optimizer.cache_audio("Hello", "MockTTS", "default", audio)
    performance_optimizer.clear_cache()
    assert performance_optimizer.get_cached_audio("Hello", "MockTTS", "default") is None

def test_get_cache_stats(performance_optimizer):
    audio1 = np.array([0.1, 0.2, 0.3])
    audio2 = np.array([0.4, 0.5, 0.6])

    performance_optimizer.cache_audio("Hello1", "MockTTS", "default", audio1)
    performance_optimizer.cache_audio("Hello2", "MockTTS", "default", audio2)

    stats = performance_optimizer.get_cache_stats()
    assert stats["cache_size"] == 2
    assert stats["max_cache_size"] == 2
    assert stats["cache_usage"] == "100.00%"

# Add more tests as needed